detect: Move aarch64 tests to common.rs
