+++ src/imp/atomic128/detect/aarch64_windows.rs
@@ -37,49 +37,3 @@ fn _detect(info: &mut CpuInfo) {
         info.set(CpuInfo::HAS_LSE);
     }
 }
-
-#[allow(clippy::undocumented_unsafe_blocks)]
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_bit_flags() {
-        let mut x = CpuInfo(0);
-        assert!(!x.test(CpuInfo::INIT));
-        assert!(!x.test(CpuInfo::HAS_LSE));
-        x.set(CpuInfo::INIT);
-        assert!(x.test(CpuInfo::INIT));
-        assert!(!x.test(CpuInfo::HAS_LSE));
-        x.set(CpuInfo::HAS_LSE);
-        assert!(x.test(CpuInfo::INIT));
-        assert!(x.test(CpuInfo::HAS_LSE));
-    }
-
-    #[test]
-    fn test_detect() {
-        if has_lse() {
-            assert!(detect().test(CpuInfo::HAS_LSE));
-            #[cfg(any(
-                target_feature = "lse",
-                portable_atomic_target_feature = "lse",
-                not(portable_atomic_no_aarch64_target_feature),
-            ))]
-            unsafe {
-                use core::{cell::UnsafeCell, sync::atomic::Ordering};
-                let v = UnsafeCell::new(0);
-                assert_eq!(
-                    super::super::_compare_exchange_casp(v.get(), 0, 1, Ordering::SeqCst),
-                    0
-                );
-                assert_eq!(*v.get(), 1);
-            }
-        } else {
-            assert!(!detect().test(CpuInfo::HAS_LSE));
-            #[cfg(not(portable_atomic_no_aarch64_target_feature))]
-            {
-                assert!(!std::arch::is_aarch64_feature_detected!("lse"));
-            }
-        }
-    }
-}
